Add ~ppc/~ppc64 to kdedglobalaccel as they was added to full kde-4.2 (Portage version: 2.2_rc23/cvs/Linux x86_64)  (Signed Manifest commit)

+++ sys-libs/db/db-4.6.21_p3-r1.ebuild
@@ -1,6 +1,6 @@
-# Copyright 1999-2009 Gentoo Foundation
+# Copyright 1999-2008 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/sys-libs/db/Attic/db-4.6.21_p3-r1.ebuild,v 1.5 2009/02/07 09:10:23 vapier Exp $
+# $Header: /var/cvsroot/gentoo-x86/sys-libs/db/Attic/db-4.6.21_p3-r1.ebuild,v 1.3 2008/09/09 05:38:29 robbat2 Exp $
 
 inherit eutils db flag-o-matic java-pkg-opt-2 autotools libtool
 
@@ -27,7 +27,7 @@ done
 LICENSE="OracleDB"
 SLOT="4.6"
 KEYWORDS="~alpha ~amd64 ~arm ~hppa ~ia64 ~m68k ~ppc ~ppc64 ~s390 ~sh ~sparc ~sparc-fbsd ~x86 ~x86-fbsd"
-IUSE="tcl java doc nocxx"
+IUSE="tcl java doc nocxx bootstrap"
 
 DEPEND="tcl? ( >=dev-lang/tcl-8.4 )
 	java? ( >=virtual/jdk-1.4 )
@@ -42,7 +42,12 @@ src_unpack() {
 	do
 		epatch "${DISTDIR}"/patch."${MY_PV}"."${i}"
 	done
+	# This patch and sed statement only matter when USE=bootstrap is in effect
+	# because the build system is regenerated otherwise.
 	epatch "${FILESDIR}"/"${PN}"-"${SLOT}"-libtool.patch
+	sed -i \
+		-e "s,\(ac_compiler\|\${MAKEFILE_CC}\|\${MAKEFILE_CXX}\|\$CC\)\( *--version\),\1 -dumpversion,g" \
+		"${S}"/../dist/configure
 
 	# use the includes from the prefix
 	epatch "${FILESDIR}"/"${PN}"-"${SLOT}"-jni-check-prefix-first.patch
@@ -54,38 +59,43 @@ src_unpack() {
 		-e '/jarfile=.*\.jar$/s,(.jar$),-$(LIBVERSION)\1,g' \
 		"${S}"/../dist/Makefile.in
 
-	cd "${S}"/../dist
-	rm -f aclocal/libtool.m4
-	sed -i \
-		-e '/AC_PROG_LIBTOOL$/aLT_OUTPUT' \
-		configure.ac
-	sed -i \
-		-e '/^AC_PATH_TOOL/s/ sh, none/ bash, none/' \
-		aclocal/programs.m4
-	AT_M4DIR="aclocal aclocal_java" eautoreconf
-	# Upstream sucks - they do autoconf and THEN replace the version variables.
-	. ./RELEASE
-	sed -i \
-		-e "s/__EDIT_DB_VERSION_MAJOR__/$DB_VERSION_MAJOR/g" \
-		-e "s/__EDIT_DB_VERSION_MINOR__/$DB_VERSION_MINOR/g" \
-		-e "s/__EDIT_DB_VERSION_PATCH__/$DB_VERSION_PATCH/g" \
-		-e "s/__EDIT_DB_VERSION_STRING__/$DB_VERSION_STRING/g" \
-		-e "s/__EDIT_DB_VERSION_UNIQUE_NAME__/$DB_VERSION_UNIQUE_NAME/g" \
-		-e "s/__EDIT_DB_VERSION__/$DB_VERSION/g" configure
+	# During bootstrap, libtool etc might not yet be available
+	if use !bootstrap; then
+		cd "${S}"/../dist
+		rm -f aclocal/libtool.m4
+		sed -i \
+			-e '/AC_PROG_LIBTOOL$/aLT_OUTPUT' \
+			configure.ac
+		sed -i \
+			-e '/^AC_PATH_TOOL/s/ sh, none/ bash, none/' \
+			aclocal/programs.m4
+		AT_M4DIR="aclocal aclocal_java" eautoreconf
+		# Upstream sucks - they do autoconf and THEN replace the version variables.
+		. ./RELEASE
+		sed -i \
+			-e "s/__EDIT_DB_VERSION_MAJOR__/$DB_VERSION_MAJOR/g" \
+			-e "s/__EDIT_DB_VERSION_MINOR__/$DB_VERSION_MINOR/g" \
+			-e "s/__EDIT_DB_VERSION_PATCH__/$DB_VERSION_PATCH/g" \
+			-e "s/__EDIT_DB_VERSION_STRING__/$DB_VERSION_STRING/g" \
+			-e "s/__EDIT_DB_VERSION_UNIQUE_NAME__/$DB_VERSION_UNIQUE_NAME/g" \
+			-e "s/__EDIT_DB_VERSION__/$DB_VERSION/g" configure
+	fi
 }
 
 src_compile() {
 	# compilation with -O0 fails on amd64, see bug #171231
 	if use amd64; then
 		replace-flags -O0 -O2
-		is-flagq -O[s123] || append-flags -O2
+		is-flag -O[s123] || append-flags -O2
 	fi
 
 	local myconf=""
 
 	use amd64 && myconf="${myconf} --with-mutex=x86/gcc-assembly"
 
-	myconf="${myconf} $(use_enable !nocxx cxx)"
+	use bootstrap \
+		&& myconf="${myconf} --disable-cxx" \
+		|| myconf="${myconf} $(use_enable !nocxx cxx)"
 
 	use tcl \
 		&& myconf="${myconf} --enable-tcl --with-tcl=/usr/$(get_libdir)" \
